refactor: remove views, multiple fixes
