+++ src/components/Modal.tsx
@@ -25,13 +25,13 @@ export function Modal({
       }}
     >
       <DialogPrimitive.Portal>
-        <DialogPrimitive.Backdrop className="fixed inset-0 z-50 bg-foreground/40 backdrop-blur-[2px] data-open:animate-in data-open:fade-in-0 data-closed:animate-out data-closed:fade-out-0" />
+        <DialogPrimitive.Backdrop className="fixed inset-0 z-50 bg-foreground/40 backdrop-1 data-open:animate-in data-open:fade-in-0 data-closed:animate-out data-closed:fade-out-0" />
         <DialogPrimitive.Popup
           className={cn(
-            "fixed top-1/2 left-1/2 z-50 flex h-[500px] max-h-[86vh] -translate-x-1/2 -translate-y-1/2 flex-col overflow-hidden rounded-2xl bg-background shadow-[0_24px_70px_color-mix(in_oklch,var(--foreground)_28%,transparent)] outline-none data-open:animate-in data-open:fade-in-0 data-open:zoom-in-95 data-closed:animate-out data-closed:fade-out-0 data-closed:zoom-out-95",
+            "fixed top-1/2 left-1/2 z-50 flex h-125 max-h-modal -translate-x-1/2 -translate-y-1/2 flex-col overflow-hidden rounded-2xl bg-background shadow-modal outline-none data-open:animate-in data-open:fade-in-0 data-open:zoom-in-95 data-closed:animate-out data-closed:fade-out-0 data-closed:zoom-out-95",
             width === "wide"
-              ? "w-full max-w-[94vw] sm:max-w-[640px]"
-              : "w-[360px] max-w-[92vw] sm:w-[420px]",
+              ? "w-full max-w-modal-wide sm:max-w-160"
+              : "w-90 max-w-modal sm:w-105",
           )}
         >
           {children}
