+++ src/index.css
@@ -34,12 +34,16 @@
   --radius-3xl: calc(var(--radius) * 2.2);
   --radius-4xl: calc(var(--radius) * 2.6);
 
-  --grid-template-columns-variety-desktop: 40px minmax(130px, 1.6fr) repeat(
-      var(--trait-cols),
-      1fr
-    ) 40px;
+  --grid-template-columns-variety-desktop: 40px minmax(130px, 1.6fr)
+    repeat(var(--trait-cols), 1fr) 40px;
   --grid-template-columns-variety-mobile: 34px minmax(0, 1.6fr) 1fr 32px;
   --grid-template-columns-yields: 40px 1fr 1fr;
+
+  --max-height-modal: 86vh;
+  --max-width-modal: 92vw;
+  --max-width-modal-wide: 94vw;
+  --shadow-modal: 0 24px 70px
+    color-mix(in oklch, var(--foreground) 28%, transparent);
 }
 
 :root {
